extract/add Representation from SignatureScheme
