+++ src/siwe_cacao.rs
@@ -1,16 +1,220 @@
-use super::{BasicSignature, Payload, SignatureScheme, VerificationError, Version};
+use super::{Representation, SignatureScheme, VerificationError, CACAO};
 use async_trait::async_trait;
+use chrono::{DateTime, Utc};
 use ethers_core::{types::H160, utils::to_checksum};
 use hex::FromHex;
-use siwe::{Message, VerificationError as SVE, Version as SVersion};
+use http::uri::{Authority, InvalidUri};
+use iri_string::{
+    types::{UriAbsoluteString, UriString},
+    validate::Error as URIStringError,
+};
+use libipld::{
+    cbor::{DagCbor, DagCborCodec},
+    codec::{Decode, Encode},
+    error::Error as IpldError,
+    DagCbor,
+};
+pub use siwe;
+use siwe::{Message, TimeStamp, VerificationError as SVE, Version as SVersion};
+use std::fmt::Debug;
 use std::io::{Read, Seek, Write};
 use thiserror::Error;
 
-use libipld::{
-    cbor::DagCborCodec,
-    codec::{Decode, Encode},
-    error::Error as IpldError,
-};
+pub type SiweCacao = CACAO<Eip191, Eip4361>;
+
+#[derive(Debug)]
+pub struct Header;
+
+#[derive(Clone, Debug)]
+pub struct Payload {
+    pub domain: Authority,
+    pub iss: UriAbsoluteString,
+    pub statement: Option<String>,
+    pub aud: UriString,
+    pub version: Version,
+    pub nonce: String,
+    pub iat: TimeStamp,
+    pub exp: Option<TimeStamp>,
+    pub nbf: Option<TimeStamp>,
+    pub request_id: Option<String>,
+    pub resources: Vec<UriString>,
+}
+
+#[derive(Copy, Clone, Debug)]
+pub enum Version {
+    V1 = 1,
+}
+
+impl Payload {
+    pub fn sign<S: SignatureScheme<Eip4361>>(self, s: S::Signature) -> CACAO<S, Eip4361>
+    where
+        S::Signature: DagCbor + Debug,
+    {
+        CACAO::new(self, s, None)
+    }
+
+    pub async fn verify<S: SignatureScheme<Eip4361>>(
+        &self,
+        s: &<S as SignatureScheme<Eip4361>>::Signature,
+    ) -> Result<(), VerificationError>
+    where
+        S: Send + Sync,
+        S::Signature: Send + Sync,
+    {
+        S::verify(self, s).await
+    }
+
+    pub fn iss(&self) -> &str {
+        self.iss.as_str()
+    }
+
+    pub fn valid_at(&self, t: &DateTime<Utc>) -> bool {
+        self.nbf.as_ref().map(|nbf| nbf < t).unwrap_or(true)
+            && self.exp.as_ref().map(|exp| exp >= t).unwrap_or(true)
+    }
+
+    pub fn valid_now(&self) -> bool {
+        self.valid_at(&Utc::now())
+    }
+}
+
+mod payload_ipld {
+    use super::*;
+    use libipld::error::Error as IpldError;
+    use std::io::{Read, Seek, Write};
+
+    #[derive(Clone, DagCbor)]
+    struct TmpPayload {
+        aud: String,
+        #[ipld(default = None)]
+        exp: Option<String>,
+        iat: String,
+        iss: String,
+        #[ipld(default = None)]
+        nbf: Option<String>,
+        nonce: String,
+        domain: String,
+        version: String,
+        resources: Vec<String>,
+        #[ipld(rename = "requestId")]
+        #[ipld(default = None)]
+        request_id: Option<String>,
+        #[ipld(default = None)]
+        statement: Option<String>,
+    }
+
+    impl From<&Payload> for TmpPayload {
+        fn from(p: &Payload) -> Self {
+            Self {
+                domain: p.domain.to_string(),
+                iss: p.iss.to_string(),
+                statement: p.statement.as_ref().map(|e| e.to_string()),
+                aud: p.aud.to_string(),
+                version: (p.version as u64).to_string(),
+                nonce: p.nonce.to_string(),
+                iat: p.iat.to_string(),
+                exp: p.exp.as_ref().map(|e| e.to_string()),
+                nbf: p.nbf.as_ref().map(|e| e.to_string()),
+                request_id: p.request_id.clone(),
+                resources: p.resources.iter().map(|r| r.to_string()).collect(),
+            }
+        }
+    }
+
+    #[derive(Error, Debug)]
+    pub enum PayloadIpldParseError {
+        #[error(transparent)]
+        Domain(#[from] InvalidUri),
+        #[error(transparent)]
+        Uri(#[from] URIStringError),
+        #[error(transparent)]
+        TimeStamp(#[from] chrono::format::ParseError),
+    }
+
+    impl TryFrom<TmpPayload> for Payload {
+        type Error = PayloadIpldParseError;
+        fn try_from(p: TmpPayload) -> Result<Self, Self::Error> {
+            Ok(Self {
+                domain: p.domain.parse()?,
+                iss: p.iss.parse()?,
+                statement: p.statement,
+                aud: p.aud.parse()?,
+                version: Version::V1,
+                nonce: p.nonce,
+                iat: p.iat.parse()?,
+                exp: p.exp.map(|s| s.parse()).transpose()?,
+                nbf: p.nbf.map(|s| s.parse()).transpose()?,
+                request_id: p.request_id,
+                resources: p
+                    .resources
+                    .iter()
+                    .map(|r| r.parse())
+                    .collect::<Result<Vec<UriString>, URIStringError>>()?,
+            })
+        }
+    }
+
+    impl Encode<DagCborCodec> for Payload {
+        fn encode<W>(&self, c: DagCborCodec, w: &mut W) -> Result<(), IpldError>
+        where
+            W: Write,
+        {
+            TmpPayload::from(self).encode(c, w)
+        }
+    }
+
+    impl Decode<DagCborCodec> for Payload {
+        fn decode<R>(c: DagCborCodec, r: &mut R) -> Result<Self, IpldError>
+        where
+            R: Read + Seek,
+        {
+            TmpPayload::decode(c, r).and_then(|t| Ok(t.try_into()?))
+        }
+    }
+
+    #[derive(DagCbor)]
+    struct DummyHeader {
+        t: String,
+    }
+    impl Encode<DagCborCodec> for Header {
+        fn encode<W>(&self, c: DagCborCodec, w: &mut W) -> Result<(), IpldError>
+        where
+            W: Write,
+        {
+            DummyHeader {
+                t: "eip4361".to_string(),
+            }
+            .encode(c, w)
+        }
+    }
+
+    #[derive(Error, Debug)]
+    #[error("Invalid Type")]
+    struct HeaderTypeErr;
+
+    impl Decode<DagCborCodec> for Header {
+        fn decode<R>(c: DagCborCodec, r: &mut R) -> Result<Self, IpldError>
+        where
+            R: Read + Seek,
+        {
+            if DummyHeader::decode(c, r)?.t != "eip4361" {
+                Err(HeaderTypeErr)?
+            } else {
+                Ok(Header)
+            }
+        }
+    }
+}
+
+pub struct Eip4361;
+
+impl Representation for Eip4361 {
+    type Payload = Payload;
+    type Header = Header;
+    fn header() -> Header {
+        Header
+    }
+}
 
 impl From<Version> for SVersion {
     fn from(s: Version) -> Self {
@@ -101,8 +305,6 @@ impl From<Message> for Payload {
 }
 
 #[derive(Debug)]
-pub struct SignInWithEthereum;
-#[derive(Debug)]
 pub struct SIWESignature([u8; 65]);
 
 impl std::ops::Deref for SIWESignature {
@@ -135,6 +337,8 @@ impl TryFrom<Vec<u8>> for SIWESignature {
 pub enum SIWESignatureDecodeError {
     #[error("Invalid length, expected 65, got {0}")]
     InvalidLength(usize),
+    #[error("Invalid Type, expected 'eip191', got {0}")]
+    InvalidType(String),
 }
 
 impl From<Vec<u8>> for SIWESignatureDecodeError {
@@ -143,12 +347,22 @@ impl From<Vec<u8>> for SIWESignatureDecodeError {
     }
 }
 
+#[derive(DagCbor)]
+struct DummySig {
+    s: Vec<u8>,
+    t: String,
+}
+
 impl Encode<DagCborCodec> for SIWESignature {
     fn encode<W>(&self, c: DagCborCodec, w: &mut W) -> Result<(), IpldError>
     where
         W: Write,
     {
-        self.0.as_ref().encode(c, w)
+        DummySig {
+            s: self.0.to_vec(),
+            t: "eip191".to_string(),
+        }
+        .encode(c, w)
     }
 }
 
@@ -157,22 +371,28 @@ impl Decode<DagCborCodec> for SIWESignature {
     where
         R: Read + Seek,
     {
-        Ok(Vec::<u8>::decode(c, r).map(SIWESignature::try_from)??)
+        match DummySig::decode(c, r)? {
+            d if d.t == "eip191" => Ok(d.s.try_into()?),
+            d => Err(SIWESignatureDecodeError::InvalidType(d.t))?,
+        }
     }
 }
 
+#[derive(Debug)]
+pub struct Eip191;
+
 #[async_trait]
-impl SignatureScheme for SignInWithEthereum {
-    type Signature = BasicSignature<SIWESignature>;
-    fn id() -> String {
-        "eip4361-eip191".into()
-    }
-    async fn verify(payload: &Payload, sig: &Self::Signature) -> Result<(), VerificationError> {
+impl SignatureScheme<Eip4361> for Eip191 {
+    type Signature = SIWESignature;
+    async fn verify(
+        payload: &<Eip4361 as Representation>::Payload,
+        sig: &Self::Signature,
+    ) -> Result<(), VerificationError> {
         let m: Message = payload
             .clone()
             .try_into()
             .map_err(|_| VerificationError::MissingVerificationMaterial)?;
-        m.verify_eip191(&sig.s.0)?;
+        m.verify_eip191(&sig)?;
         Ok(())
     }
 }
@@ -202,12 +422,12 @@ Issued At: 2021-12-07T18:28:18.807Z"#,
         .unwrap()
         .into();
         // correct signature
-        SignInWithEthereum::verify(&message, &<Vec<u8>>::from_hex(r#"6228b3ecd7bf2df018183aeab6b6f1db1e9f4e3cbe24560404112e25363540eb679934908143224d746bbb5e1aa65ab435684081f4dbb74a0fec57f98f40f5051c"#).unwrap().try_into().unwrap())
+        Eip191::verify(&message, &<Vec<u8>>::from_hex(r#"6228b3ecd7bf2df018183aeab6b6f1db1e9f4e3cbe24560404112e25363540eb679934908143224d746bbb5e1aa65ab435684081f4dbb74a0fec57f98f40f5051c"#).unwrap().try_into().unwrap())
             .await
             .unwrap();
 
         // incorrect signature
-        assert!(SignInWithEthereum::verify(&message, &<Vec<u8>>::from_hex(r#"7228b3ecd7bf2df018183aeab6b6f1db1e9f4e3cbe24560404112e25363540eb679934908143224d746bbb5e1aa65ab435684081f4dbb74a0fec57f98f40f5051c"#).unwrap().try_into().unwrap())
+        assert!(Eip191::verify(&message, &<Vec<u8>>::from_hex(r#"7228b3ecd7bf2df018183aeab6b6f1db1e9f4e3cbe24560404112e25363540eb679934908143224d746bbb5e1aa65ab435684081f4dbb74a0fec57f98f40f5051c"#).unwrap().try_into().unwrap())
             .await
             .is_err());
     }
